kmodsrc: support RONX .text
When CONFIG_DEBUG_SET_MODULE_RONX is set, the text of modules is not
writable and patching such memory causes an oops. Fix that by marking
.text RW temporarily during patching.

Note that we cannot call the function that marks .text RW before we
bootstrap as it is not exported and EXTRACT_SYMBOL does not work in
bootstrap phase yet. So we instead find that one specific function and
use that.

Signed-off-by: Jiri Slaby <jslaby@suse.cz>