fix set_all_modules_text_rw fallout
When CONFIG_DEBUG_SET_MODULE_RONX is not set, set_all_modules_text_rw
is an inline function. So we are trying to find address of a ghost. Do
that only if DEBUG_SET_MODULE_RONX is set.

This is a regression reported by Richard and which I introduced in
59647168dd7cab3e199b9c11f2bf2f2d89f46900 (kmodsrc: support RONX
.text).

Signed-off-by: Jiri Slaby <jslaby@suse.cz>
Reported-and-Tested-by: Richard Weinberger <richard@nod.at>